Separate $schedule data sources and add Roadmap tab (#148)
* Move resources into directories

* Split schedule.json into many files

* Update tsconfig.json

Add resolveJsonModule

* Begin migrating to use requires

* Update schedule.json

Add the January 2021 schedule

* Update schedule.json

* Update schedule.json

* Update schedule.json

Added end of January event and the next Guild Wars

* Update schedule.json

* Allow users to directly query upcoming schedule (#147)

* Update festival.json

* Update festival.json

* Move resources into directories

* Split schedule.json into many files

* Remove logs

* Move features.json to roadmap.json

* Add most recent events

* Add parsing for roadmap.json

* Update JSON format

* Fix selecting a page

Previously, we were only able to select the page in the 1th position due to a bug. It has been fixed.

* Update schedule.ts

Here, we update a bunch of the data pipeline methods to read from separate files. We also enabled the roadmap and calling it directly.

* Delete schedule.json